New Soundfiler. Please read carefully the detailed description.
Issue #893
The module still contain a bug, but I don't know where exactly it is.

To reproduce it:
- Drag a Soundfiler;
- Record a sound;
- Play the recorded sound without stop the recording session. (If you press "Stop" before to press "Play", will happen the same thing)

In all other usage it work totally.
- If you press "Pause" and then "Play" during the reading process or you try to change the reading position in different regions of the audio file, you are not able to listen any clicks.
- With loaded files it totaly works.
- If you test the module inside the Module Creator (using just Pd), the soundfiler works well

BUT

when you will use it in IntegraLive you can listen some some clicks, but just at the beginning of the file.

Maybe the bugs source is in the communication between Pd and IntegraLive; and it is generated only after that you have recorded a sound file and then you wanna play it.
Maybe is a message sent from Pd to IntegraLive, maybe is the refreshing of the position value. I can't understand very well actually.

Please test the module and let me know your opinion about it.

---

Below the Pd report of the following actions:
ACTIONS:
- Drag a soundfiler;
- Record a sound;
- Play the recorded sound without stop the recording session.

REPORT (from the Pd window):
Created new readanysf~ with 1 channels and internal buffer of 32 blocks of 256 samples = 8192
sample-1234: no such table
sample-1234: no such table
readanysf~: looping = 0
can't seek on this file.
readanysf~: Current file is still starting.
This probably means that it is a stream and it needs to buffer in from the network.
readanysf~: Invalid file or unsupported codec.
can't seek on this file.
readanysf~: Current file is either invalid or an unsupported codec.
can't seek on this file.
readanysf~: Current file is still starting.
This probably means that it is a stream and it needs to buffer in from the network.
readanysf~: Current file is still starting.
This probably means that it is a stream and it needs to buffer in from the network.