Reinstate "core" tag for Audio IO modules
"core" is a reserved tag, which carries a special signifcance within the Integra framework. Specifically:

-  "core" modules are guaranteed to exist, so customers of the API can assume that they will be available
- they are guaranteed to be maintained only by Integra developers
- there are only ever shipped with Integra Live, they can't be installed as a third party module

"core" is a contract with customer code, which makes it "safe"  to make certain assumptions about these modules.

If we don't want the "core" tag to show up in the Integra Live module tag filter, we need a solution other than just removing the tag from the interface definitions, as this simply breaks the software (e.g. it hangs at startup)