Ensure JuceGUITest2 compiles and links in place
The original project was in a different location, so header paths etc. were wrong. I also modified the .jucer project to ensure the .xcodeproj could be rebuilt. The only thing missing after a Projucer rebuild is to set the file libmarkdownparser.a membership checkbox to "required" for the JuceGUITest2 app.